second commit - minor updates

+++ Inventory_system/Inventory_system/Controllers/ProductController.cs
@@ -9,46 +9,85 @@ namespace Inventory_system.Controllers
     public class ProductController : Controller
     {
         private IProductService _service;
+
         public ProductController(IProductService service)
         {
             _service = service;
+            //logging - to track errors and events
         }   
 
 
-        //To add a new product with variants and sub-variants
-        [Route("CreateProduct")]
-        [HttpPost]
+        //Create new product with variants and sub-variants
+        //returns 200 Ok if successful
+        [HttpPost("CreateProduct")]
         public IActionResult CreateProduct([FromBody] ProductDto product)
         {
+            try
+            {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             _service.CreateProduct(product);
             return Ok();
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, $"An error occurred while processing your request.{ex.Message}");
+            }
         }
 
         //To add stock quantity (purchase)
-        [Route("AddStock")]
-        [HttpPost]
+        [HttpPost("AddStock")]
         public IActionResult AddStock(StockRequestDto AddStock)
         {
+            try
+            {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             _service.AddStock(AddStock);
             return Ok();
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, $"An error occurred while processing your request.{ex.Message}");
+            }
         }
 
         //To remove or decrease stock quantity (sales)
-        [Route("RemoveStock")]
-        [HttpPost]
+        [HttpPost("RemoveStock")]
         public IActionResult RemoveStock([FromBody] StockRequestDto RemoveStock)
         {
+            try
+            {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             _service.RemoveStock(RemoveStock);
             return Ok();
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, $"An error occurred while processing your request.{ex.Message}");
+            }
         }
 
         //To list available products with variants and sub-variants
-        [Route("Products")]
-        [HttpGet]
+        [HttpGet("Products")]
         public IActionResult ListProducts()
         {
+            try
+            {
             var products = _service.ListProducts();
             return Ok(products);
+            }
+            catch (Exception ex)
+            {
+                return StatusCode(500, $"An error occurred while processing your request.{ex.Message}");
+            }
         }
     }
 }
